+++ test/e2e/specs/specs-wpcom/wp-calypso-gutenberg-site-editor-tracking-spec.js
@@ -14,8 +14,8 @@ import SiteEditorComponent from '../../../lib/components/site-editor-component.j
 
 import * as driverManager from '../../../lib/driver-manager.js';
 import * as dataHelper from '../../../lib/data-helper.js';
-import { clearEventsStack } from './utils.js';
-import { createGeneralTests } from './general-tests.js';
+import { clearEventsStack } from '../../lib/gutenberg/tracking/utils.js';
+import { createGeneralTests } from '../../lib/gutenberg/tracking/general-tests.js';
 
 const mochaTimeOut = config.get( 'mochaTimeoutMS' );
 const screenSize = driverManager.currentScreenSize();
